+++ client/lib/page_cluster.js
@@ -1,8 +1,21 @@
 "use strict";
 /*jslint todo: true, regexp: true, browser: true, unparam: true, plusplus: true */
 /*global Promise */
+var State = require('./state.js');
 var PageTable = require('./page_table.js');
 var DisplayError = require('./alerts.js').prototype.DisplayError;
+
+/* Clusters should link back to an equivalent concordance */
+function renderCluster(url_prefix, data, type, full, meta) {
+    if (type === 'display') {
+        return '<a title="Click to find individual concordances" target="_blank"' +
+               ' onclick="event.stopPropagation();"' +
+               ' href="' + url_prefix + '&conc-q=' + encodeURIComponent(data) + '"' +
+               '>' + data + '</a>';
+    }
+
+    return data;
+}
 
 // PageCluster inherits PageTable
 function PageCluster() {
@@ -17,7 +30,7 @@ PageCluster.prototype.init = function () {
     this.table_opts.autoWidth = false;
     this.table_opts.columns = [
         { title: "", defaultContent: "", width: "3rem", sortable: false, searchable: false },
-        { title: "Cluster", data: "0"},
+        { title: "Cluster", data: "0", render: renderCluster },
         { title: "Frequency", data: "1"},
     ];
     this.table_opts.order = [[2, "desc"]];
@@ -26,6 +39,19 @@ PageCluster.prototype.init = function () {
 
 PageCluster.prototype.page_title = function (page_state) {
     return "CLiC clusters search";
+};
+
+PageCluster.prototype.reload = function reload(page_state) {
+    var new_state = new State(window, page_state.defaults);
+
+    // Make a URL pointing at the concordance page, with the same corpora, use this in cluster links
+    new_state.update({doc: 'concordance', args: {
+        corpora: page_state.arg('corpora'),
+        'conc-subset': page_state.arg('subset'),
+    }}, true);
+    this.table_opts.columns[1].render = renderCluster.bind(null, new_state.to_url());
+
+    return PageTable.prototype.reload.apply(this, arguments);
 };
 
 PageCluster.prototype.reload_data = function reload(page_state) {