+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Link cluster lines to a matching concordance search
 * Update page title on page changes
 * Make error/warning/info reporting more consistent
 * Add basic twitter, opengraph preview support